Move component test utilities to src/test-util/
Improve organizational clarity by separating tests from utility modules
used in tests, and also pave the way for re-using these utilities for UI
outside the sidebar app in future (eg. src/annotator/).
