+++ src/sidebar/components/test/annotation-body-test.js
@@ -7,8 +7,8 @@ import * as fixtures from '../../test/annotation-fixtures';
 import AnnotationBody from '../annotation-body';
 import { $imports } from '../annotation-body';
 
-import { checkAccessibility } from './accessibility';
-import mockImportedComponents from './mock-imported-components';
+import { checkAccessibility } from '../../../test-util/accessibility';
+import mockImportedComponents from '../../../test-util/mock-imported-components';
 
 describe('AnnotationBody', () => {
   function createBody(props = {}) {
