+++ src/sidebar/components/test/annotation-publish-control-test.js
@@ -5,8 +5,8 @@ import * as fixtures from '../../test/annotation-fixtures';
 import AnnotationPublishControl from '../annotation-publish-control';
 import { $imports } from '../annotation-publish-control';
 
-import { checkAccessibility } from './accessibility';
-import mockImportedComponents from './mock-imported-components';
+import { checkAccessibility } from '../../../test-util/accessibility';
+import mockImportedComponents from '../../../test-util/mock-imported-components';
 
 describe('AnnotationPublishControl', () => {
   let fakeGroup;
