+++ src/sidebar/components/test/annotation-user-test.js
@@ -4,8 +4,8 @@ import { createElement } from 'preact';
 import AnnotationUser from '../annotation-user';
 import { $imports } from '../annotation-user';
 
-import { checkAccessibility } from './accessibility';
-import mockImportedComponents from './mock-imported-components';
+import { checkAccessibility } from '../../../test-util/accessibility';
+import mockImportedComponents from '../../../test-util/mock-imported-components';
 
 describe('AnnotationUser', () => {
   let fakeAnnotation;
