+++ src/sidebar/components/test/button-test.js
@@ -4,8 +4,8 @@ import { createElement } from 'preact';
 import Button from '../button';
 import { $imports } from '../button';
 
-import { checkAccessibility } from './accessibility';
-import mockImportedComponents from './mock-imported-components';
+import { checkAccessibility } from '../../../test-util/accessibility';
+import mockImportedComponents from '../../../test-util/mock-imported-components';
 
 describe('Button', () => {
   let fakeOnClick;
