+++ src/sidebar/components/test/focused-mode-header-test.js
@@ -4,8 +4,8 @@ import { createElement } from 'preact';
 import FocusedModeHeader from '../focused-mode-header';
 import { $imports } from '../focused-mode-header';
 
-import { checkAccessibility } from './accessibility';
-import mockImportedComponents from './mock-imported-components';
+import { checkAccessibility } from '../../../test-util/accessibility';
+import mockImportedComponents from '../../../test-util/mock-imported-components';
 
 describe('FocusedModeHeader', function() {
   let fakeStore;
