+++ src/sidebar/components/test/logged-out-message-test.js
@@ -4,8 +4,8 @@ import { createElement } from 'preact';
 import LoggedOutMessage from '../logged-out-message';
 import { $imports } from '../logged-out-message';
 
-import { checkAccessibility } from './accessibility';
-import mockImportedComponents from './mock-imported-components';
+import { checkAccessibility } from '../../../test-util/accessibility';
+import mockImportedComponents from '../../../test-util/mock-imported-components';
 
 describe('LoggedOutMessage', () => {
   const createLoggedOutMessage = props => {
