+++ src/sidebar/components/test/search-status-bar-test.js
@@ -4,8 +4,8 @@ import { createElement } from 'preact';
 import SearchStatusBar from '../search-status-bar';
 import { $imports } from '../search-status-bar';
 
-import { checkAccessibility } from './accessibility';
-import mockImportedComponents from './mock-imported-components';
+import { checkAccessibility } from '../../../test-util/accessibility';
+import mockImportedComponents from '../../../test-util/mock-imported-components';
 
 describe('SearchStatusBar', () => {
   let fakeRootThread;
