+++ src/sidebar/components/test/share-annotations-panel-test.js
@@ -4,8 +4,8 @@ import { createElement } from 'preact';
 import ShareAnnotationsPanel from '../share-annotations-panel';
 import { $imports } from '../share-annotations-panel';
 
-import { checkAccessibility } from './accessibility';
-import mockImportedComponents from './mock-imported-components';
+import { checkAccessibility } from '../../../test-util/accessibility';
+import mockImportedComponents from '../../../test-util/mock-imported-components';
 
 describe('ShareAnnotationsPanel', () => {
   let fakeStore;
