+++ src/sidebar/components/test/share-links-test.js
@@ -4,8 +4,8 @@ import { createElement } from 'preact';
 import ShareLinks from '../share-links';
 import { $imports } from '../share-links';
 
-import { checkAccessibility } from './accessibility';
-import mockImportedComponents from './mock-imported-components';
+import { checkAccessibility } from '../../../test-util/accessibility';
+import mockImportedComponents from '../../../test-util/mock-imported-components';
 
 describe('ShareLinks', () => {
   let fakeAnalytics;
