+++ src/sidebar/components/test/sidebar-panel-test.js
@@ -4,8 +4,8 @@ import { createElement } from 'preact';
 import SidebarPanel from '../sidebar-panel';
 import { $imports } from '../sidebar-panel';
 
-import { checkAccessibility } from './accessibility';
-import mockImportedComponents from './mock-imported-components';
+import { checkAccessibility } from '../../../test-util/accessibility';
+import mockImportedComponents from '../../../test-util/mock-imported-components';
 
 describe('SidebarPanel', () => {
   let fakeStore;
