+++ src/test-util/mock-imported-components.js
